move utils to utils folder
