+++ src/utils/random-number.ts
@@ -1,0 +1,8 @@
+export default function getRandomNumberBetween(
+  min: number,
+  max: number
+): number {
+  min = Math.ceil(min);
+  max = Math.floor(max);
+  return Math.floor(Math.random() * (max - min + 1) + min);
+}
